+++ src/is-associativearray.ts
@@ -1,8 +1,7 @@
+import { AssociativeArray } from './associativearray';
 import { not } from 'existjs';
 
-// Unfortunately 'object.constructor.name' may doesn't work
-export function isAssociativeArray(object: any) {
+export function isAssociativeArray(object: AssociativeArray) {
   if (not(object)) return false;
-  if (not(object.type)) return false;
-  return (object.type === "AssociativeArray") ? true : false;
+  return (object instanceof AssociativeArray) ? true : false;
 }